# docs/openapi.yaml
openapi: 3.1.0
info:
  title: network-debugger API
  version: 0.1.0-mvp
servers:
  - url: http://localhost:9092
paths:
...
  /_api/v1/compose/library/requests:
    post:
      summary: Upsert request template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComposeRequestTemplate'
      responses:
        '200': { description: OK }
    put:
      summary: Upsert request template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComposeRequestTemplate'
      responses:
        '200': { description: OK }
...
components:
  schemas:
...
    ComposeHeader:
      type: object
      properties: { key: {type: string}, value: {type: string} }
    ComposeQuery:
      type: object
      properties: { key: {type: string}, value: {type: string} }
    ComposeFormField:
      type: object
      properties: { key: {type: string}, value: {type: string} }
    ComposePart:
      type: object
      properties:
        name: { type: string }
        isFile: { type: boolean }
        fileName: { type: string }
        contentType: { type: string }
        base64: { type: string }
        value: { type: string }
    ComposeAuth:
      type: object
      properties:
        type: { type: string, enum: [none, basic, bearer, apiKey] }
        username: { type: string }
        password: { type: string }
        bearerToken: { type: string }
        apiKey: { type: string }
        apiKeyHeader: { type: string }
    ComposeBody:
      type: object
      properties:
        mode: { type: string, enum: [raw, json, form, multipart] }
        raw: { type: string }
        json: { type: string }
        form:
          type: array
          items: { $ref: '#/components/schemas/ComposeFormField' }
        multipart:
          type: array
          items: { $ref: '#/components/schemas/ComposePart' }
    ComposeRequestTemplate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        method: { type: string }
        url: { type: string }
        headers:
          type: array
          items: { $ref: '#/components/schemas/ComposeHeader' }
        query:
          type: array
          items: { $ref: '#/components/schemas/ComposeQuery' }
        body: { $ref: '#/components/schemas/ComposeBody' }
        auth: { $ref: '#/components/schemas/ComposeAuth' }
...